+++ packages/blade/src/components/Table/TableContext.tsx
@@ -2,7 +2,7 @@
 // eslint-disable-next-line @typescript-eslint/no-empty-function
 import React from 'react';
 import type { TableNode } from '@table-library/react-table-library/table';
-import type { TableProps } from './types';
+import type { TableProps, TablePaginationType } from './types';
 import type { SurfaceLevels } from '~tokens/theme/theme';
 
 export type TableContextType = {
@@ -29,6 +29,8 @@ export type TableContextType = {
   surfaceLevel: SurfaceLevels;
   disabledRows: TableNode['id'][];
   setDisabledRows: React.Dispatch<React.SetStateAction<TableNode['id'][]>>;
+  paginationType: NonNullable<TablePaginationType>;
+  setPaginationType: React.Dispatch<React.SetStateAction<NonNullable<TablePaginationType>>>;
 };
 
 const TableContext = React.createContext<TableContextType>({
@@ -49,6 +51,8 @@ const TableContext = React.createContext<TableContextType>({
   surfaceLevel: 1,
   disabledRows: [],
   setDisabledRows: () => {},
+  paginationType: 'client',
+  setPaginationType: () => {},
 });
 
 const useTableContext = (): TableContextType => {